Rename Domain_*_Set commands to Domain_Set_*
