+++ test/lib/mock/response-data.php
@@ -113,7 +113,7 @@ function get_mock_response( Command\Command_Interface $command, ?string $domain,
 			];
 			break;
 
-		case 'Domain_Contacts_Set-success':
+		case 'Domain_Set_Contacts-success':
 			$response = [
 				'status' => 200,
 				'status_description' => 'Command completed successfully',
